backend/requirements.txt
sqlalchemy
fastapi
uvicorn[standard]
pydantic
pydantic-settings
psycopg2-binary
python-jose[cryptography]
passlib[bcrypt]
python-multipart
reportlab
svglib
pillow
email-validator
httpx
pyyaml
pytest
pytest-asyncio
